@@ -4,15 +4,13 @@
   "description": "A minimal node-pty Electron example",
   "main": "main.js",
   "scripts": {
-    "start": "electron ."
+    "start": "electron . --ignore-gpu-blacklist"
   },
   "repository": "https://github.com/Tyriar/node-pty",
   "author": "Tyriar",
-  "devDependencies": {
-    "electron": "1.6.2"
-  },
   "dependencies": {
-    "node-pty": "^0.6.3",
-    "xterm": "^2.4.0"
+    "electron": "4.0.1",
+    "node-pty": "^0.9.0",
+    "xterm": "4.1.0"
   }
 }
