@@ -2,12 +2,12 @@
   "name": "node-pty",
   "description": "Fork pseudoterminals in Node.JS",
   "author": {
-    "name": "Daniel Imms"
+    "name": "Microsoft Corporation"
   },
-  "version": "0.7.2",
+  "version": "0.9.0",
   "license": "MIT",
   "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
+  "types": "./typings/node-pty.d.ts",
   "repository": {
     "type": "git",
     "url": "git://github.com/Tyriar/node-pty.git"
@@ -17,7 +17,8 @@
     "lib/",
     "scripts/",
     "src/",
-    "deps/"
+    "deps/",
+    "typings/"
   ],
   "homepage": "https://github.com/Tyriar/node-pty",
   "bugs": {
@@ -33,22 +34,25 @@
   ],
   "scripts": {
     "tsc": "tsc",
-    "tslint": "tslint src/**/*.ts",
+    "watch": "tsc -w",
+    "lint": "tslint 'src/**/*.ts'",
     "install": "node scripts/install.js",
     "postinstall": "node scripts/post-install.js",
-    "test": "cross-env NODE_ENV=test mocha -R spec lib/*.test.js",
-    "prepublish": "npm run tsc"
+    "test": "cross-env NODE_ENV=test mocha -R spec --exit lib/*.test.js",
+    "prepare": "npm run tsc",
+    "prepublishOnly": "npm run tsc"
   },
   "dependencies": {
-    "nan": "^2.6.2"
+    "nan": "^2.14.0"
   },
   "devDependencies": {
-    "@types/mocha": "^2.2.41",
-    "@types/node": "^6.0.58",
-    "cross-env": "^3.2.4",
-    "mocha": "^3.1.2",
-    "pollUntil": "^1.0.3",
-    "tslint": "^4.3.1",
-    "typescript": "^2.1.4"
+    "@types/mocha": "^5.0.0",
+    "@types/node": "8",
+    "cross-env": "^5.1.4",
+    "mocha": "^5.0.5",
+    "ps-list": "^6.0.0",
+    "tslint": "^5.12.1",
+    "tslint-consistent-codestyle": "^1.15.0",
+    "typescript": "3.4"
   }
 }
