@@ -1,6 +1,6 @@
 var os = require('os');
 var pty = require('node-pty');
-var Terminal = require('xterm');
+var Terminal = require('xterm').Terminal;
 
 // Initialize node-pty with an appropriate shell
 const shell = process.env[os.platform() === 'win32' ? 'COMSPEC' : 'SHELL'];
@@ -17,9 +17,7 @@ const xterm = new Terminal();
 xterm.open(document.getElementById('xterm'));
 
 // Setup communication between xterm.js and node-pty
-xterm.on('data', (data) => {
-  ptyProcess.write(data);
-});
+xterm.onData(data => ptyProcess.write(data));
 ptyProcess.on('data', function (data) {
   xterm.write(data);
 });
