@@ -34,15 +34,7 @@
 #if defined(__GLIBC__) || defined(__CYGWIN__)
 #include <pty.h>
 #elif defined(__APPLE__) || defined(__OpenBSD__) || defined(__NetBSD__)
-/**
- * From node v0.10.28 (at least?) there is also a "util.h" in node/src, which
- * would confuse the compiler when looking for "util.h".
- */
-#if NODE_VERSION_AT_LEAST(0, 10, 28)
-#include <../include/util.h>
-#else
 #include <util.h>
-#endif
 #elif defined(__FreeBSD__)
 #include <libutil.h>
 #elif defined(__sun)
@@ -52,6 +44,14 @@
 #endif
 
 #include <termios.h> /* tcgetattr, tty_ioctl */
+
+/* Some platforms name VWERASE and VDISCARD differently */
+#if !defined(VWERASE) && defined(VWERSE)
+#define VWERASE	VWERSE
+#endif
+#if !defined(VDISCARD) && defined(VDISCRD)
+#define VDISCARD	VDISCRD
+#endif
 
 /* environ for execvpe */
 /* node/src/node_child_process.cc */
@@ -120,11 +120,7 @@ static void
 pty_waitpid(void *);
 
 static void
-#if NODE_VERSION_AT_LEAST(0, 11, 0)
 pty_after_waitpid(uv_async_t *);
-#else
-pty_after_waitpid(uv_async_t *, int);
-#endif
 
 static void
 pty_after_close(uv_handle_t *);
@@ -151,7 +147,7 @@ NAN_METHOD(PtyFork) {
   signal(SIGINT, SIG_DFL);
 
   // file
-  v8::String::Utf8Value file(info[0]->ToString());
+  Nan::Utf8String file(info[0]);
 
   // args
   int i = 0;
@@ -162,7 +158,7 @@ NAN_METHOD(PtyFork) {
   argv[0] = strdup(*file);
   argv[argl-1] = NULL;
   for (; i < argc; i++) {
-    v8::String::Utf8Value arg(argv_->Get(Nan::New<v8::Integer>(i))->ToString());
+    Nan::Utf8String arg(Nan::Get(argv_, i).ToLocalChecked());
     argv[i+1] = strdup(*arg);
   }
 
@@ -173,18 +169,18 @@ NAN_METHOD(PtyFork) {
   char **env = new char*[envc+1];
   env[envc] = NULL;
   for (; i < envc; i++) {
-    v8::String::Utf8Value pair(env_->Get(Nan::New<v8::Integer>(i))->ToString());
+    Nan::Utf8String pair(Nan::Get(env_, i).ToLocalChecked());
     env[i] = strdup(*pair);
   }
 
   // cwd
-  v8::String::Utf8Value cwd_(info[3]->ToString());
+  Nan::Utf8String cwd_(info[3]);
   char *cwd = strdup(*cwd_);
 
   // size
   struct winsize winp;
-  winp.ws_col = info[4]->IntegerValue();
-  winp.ws_row = info[5]->IntegerValue();
+  winp.ws_col = info[4]->IntegerValue(Nan::GetCurrentContext()).FromJust();
+  winp.ws_row = info[5]->IntegerValue(Nan::GetCurrentContext()).FromJust();
   winp.ws_xpixel = 0;
   winp.ws_ypixel = 0;
 
@@ -192,7 +188,7 @@ NAN_METHOD(PtyFork) {
   struct termios t = termios();
   struct termios *term = &t;
   term->c_iflag = ICRNL | IXON | IXANY | IMAXBEL | BRKINT;
-  if (info[8]->ToBoolean()->Value()) {
+  if (Nan::To<bool>(info[8]).FromJust()) {
 #if defined(IUTF8)
     term->c_iflag |= IUTF8;
 #endif
@@ -227,8 +223,8 @@ NAN_METHOD(PtyFork) {
   cfsetospeed(term, B38400);
 
   // uid / gid
-  int uid = info[6]->IntegerValue();
-  int gid = info[7]->IntegerValue();
+  int uid = info[6]->IntegerValue(Nan::GetCurrentContext()).FromJust();
+  int gid = info[7]->IntegerValue(Nan::GetCurrentContext()).FromJust();
 
   // fork the pty
   int master = -1;
@@ -312,8 +308,8 @@ NAN_METHOD(PtyOpen) {
 
   // size
   struct winsize winp;
-  winp.ws_col = info[0]->IntegerValue();
-  winp.ws_row = info[1]->IntegerValue();
+  winp.ws_col = info[0]->IntegerValue(Nan::GetCurrentContext()).FromJust();
+  winp.ws_row = info[1]->IntegerValue(Nan::GetCurrentContext()).FromJust();
   winp.ws_xpixel = 0;
   winp.ws_ypixel = 0;
 
@@ -357,16 +353,22 @@ NAN_METHOD(PtyResize) {
     return Nan::ThrowError("Usage: pty.resize(fd, cols, rows)");
   }
 
-  int fd = info[0]->IntegerValue();
+  int fd = info[0]->IntegerValue(Nan::GetCurrentContext()).FromJust();
 
   struct winsize winp;
-  winp.ws_col = info[1]->IntegerValue();
-  winp.ws_row = info[2]->IntegerValue();
+  winp.ws_col = info[1]->IntegerValue(Nan::GetCurrentContext()).FromJust();
+  winp.ws_row = info[2]->IntegerValue(Nan::GetCurrentContext()).FromJust();
   winp.ws_xpixel = 0;
   winp.ws_ypixel = 0;
 
   if (ioctl(fd, TIOCSWINSZ, &winp) == -1) {
-    return Nan::ThrowError("ioctl(2) failed.");
+    switch (errno) {
+      case EBADF: return Nan::ThrowError("ioctl(2) failed, EBADF");
+      case EFAULT: return Nan::ThrowError("ioctl(2) failed, EFAULT");
+      case EINVAL: return Nan::ThrowError("ioctl(2) failed, EINVAL");
+      case ENOTTY: return Nan::ThrowError("ioctl(2) failed, ENOTTY");
+    }
+    return Nan::ThrowError("ioctl(2) failed");
   }
 
   return info.GetReturnValue().SetUndefined();
@@ -384,9 +386,9 @@ NAN_METHOD(PtyGetProc) {
     return Nan::ThrowError("Usage: pty.process(fd, tty)");
   }
 
-  int fd = info[0]->IntegerValue();
+  int fd = info[0]->IntegerValue(Nan::GetCurrentContext()).FromJust();
 
-  v8::String::Utf8Value tty_(info[1]->ToString());
+  Nan::Utf8String tty_(info[1]);
   char *tty = strdup(*tty_);
   char *name = pty_getproc(fd, tty);
   free(tty);
@@ -470,11 +472,7 @@ pty_waitpid(void *data) {
  */
 
 static void
-#if NODE_VERSION_AT_LEAST(0, 11, 0)
 pty_after_waitpid(uv_async_t *async) {
-#else
-pty_after_waitpid(uv_async_t *async, int unhelpful) {
-#endif
   Nan::HandleScope scope;
   pty_baton *baton = static_cast<pty_baton*>(async->data);
 
@@ -486,7 +484,8 @@ pty_after_waitpid(uv_async_t *async, int unhelpful) {
   v8::Local<v8::Function> cb = Nan::New<v8::Function>(baton->cb);
   baton->cb.Reset();
   memset(&baton->cb, -1, sizeof(baton->cb));
-  Nan::Callback(cb).Call(Nan::GetCurrentContext()->Global(), 2, argv);
+  Nan::AsyncResource resource("pty_after_waitpid");
+  resource.runInAsyncScope(Nan::GetCurrentContext()->Global(), cb, 2, argv);
 
   uv_close((uv_handle_t *)async, pty_after_close);
 }
@@ -700,18 +699,10 @@ pty_forkpty(int *amaster,
 
 NAN_MODULE_INIT(init) {
   Nan::HandleScope scope;
-  Nan::Set(target,
-           Nan::New<v8::String>("fork").ToLocalChecked(),
-           Nan::New<v8::FunctionTemplate>(PtyFork)->GetFunction());
-  Nan::Set(target,
-           Nan::New<v8::String>("open").ToLocalChecked(),
-           Nan::New<v8::FunctionTemplate>(PtyOpen)->GetFunction());
-  Nan::Set(target,
-           Nan::New<v8::String>("resize").ToLocalChecked(),
-           Nan::New<v8::FunctionTemplate>(PtyResize)->GetFunction());
-  Nan::Set(target,
-           Nan::New<v8::String>("process").ToLocalChecked(),
-           Nan::New<v8::FunctionTemplate>(PtyGetProc)->GetFunction());
+  Nan::Export(target, "fork", PtyFork);
+  Nan::Export(target, "open", PtyOpen);
+  Nan::Export(target, "resize", PtyResize);
+  Nan::Export(target, "process", PtyGetProc);
 }
 
 NODE_MODULE(pty, init)
