@@ -1,17 +1,17 @@
 /**
  * Copyright (c) 2013-2015, Christopher Jeffrey, Peter Sunde (MIT License)
  * Copyright (c) 2016, Daniel Imms (MIT License).
+ * Copyright (c) 2018, Microsoft Corporation (MIT License).
  */
 
 #include <nan.h>
 #include <Shlwapi.h> // PathCombine
-#include <winpty.h>
 
 #include "path_util.h"
 
 namespace path_util {
 
-const wchar_t* to_wstring(const v8::String::Utf8Value& str) {
+const wchar_t* to_wstring(const Nan::Utf8String& str) {
   const char *bytes = *str;
   unsigned int sizeOfStr = MultiByteToWideChar(CP_UTF8, 0, bytes, -1, NULL, 0);
   wchar_t *output = new wchar_t[sizeOfStr];
