@@ -1,6 +1,7 @@
 /**
  * Copyright (c) 2013-2015, Christopher Jeffrey, Peter Sunde (MIT License)
  * Copyright (c) 2016, Daniel Imms (MIT License).
+ * Copyright (c) 2018, Microsoft Corporation (MIT License).
  */
 
 #ifndef NODE_PTY_PATH_UTIL_H_
@@ -12,7 +13,7 @@
 
 namespace path_util {
 
-const wchar_t* to_wstring(const v8::String::Utf8Value& str);
+const wchar_t* to_wstring(const Nan::Utf8String& str);
 bool file_exists(std::wstring filename);
 std::wstring get_shell_path(std::wstring filename);
 
